# docs/guides/2026-03-10-支撑系统接口文档-openapi3.1.yaml
...
openapi: 3.1.0
info:
  title: EleTender 支撑系统接口文档
  description: 基于运行中 ele-tender-support 服务自动生成，并补充关键枚举与中文说明。
  contact:
    name: EleTender
  version: 1.0.0
servers:
- url: http://127.0.0.1:8080
  description: Generated server url
paths:
  /api/access-logs:
    get:
      tags:
      - 访问日志管理
      summary: 分页查询访问日志
      operationId: getList_4
      parameters:
      - name: pageNum
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: traceId
        in: query
        required: false
        schema:
          type: string
      - name: serviceName
        in: query
        required: false
        schema:
          type: string
      - name: statusCode
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: bizType
        in: query
        required: false
        schema:
          type: string
      - name: bizId
        in: query
        required: false
        schema:
          type: string
      - name: projectId
        in: query
        required: false
        schema:
          type: string
      - name: tenderId
        in: query
        required: false
        schema:
          type: string
      - name: startTime
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: endTime
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageSysAccessLog'
...
  /api/external-systems:
    get:
      tags:
      - 接入系统管理
      summary: 分页查询接入系统
      operationId: getList_3
      parameters:
      - name: pageNum
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: systemName
        in: query
        required: false
        schema:
          type: string
      - name: appKey
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageSysAccessSystem'
    post:
      tags:
      - 接入系统管理
      summary: 创建接入系统
      operationId: create_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysAccessSystem'
        required: true
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultSysAccessSystem'
...
  /api/external-systems/{id}/status:
    put:
      tags:
      - 接入系统管理
      summary: 启用/禁用接入系统
      operationId: changeStatus_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultVoid'
...
  /api/roles:
    get:
      tags:
      - 角色管理
      summary: 分页查询角色
      operationId: getList_2
      parameters:
      - name: pageNum
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: roleName
        in: query
        required: false
        schema:
          type: string
      - name: roleCode
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageSysRole'
    post:
      tags:
      - 角色管理
      summary: 创建角色
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysRole'
        required: true
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultSysRole'
...
  /api/users:
    get:
      tags:
      - 用户管理
      summary: 分页查询用户
      operationId: getList_1
      parameters:
      - name: pageNum
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: username
        in: query
        required: false
        schema:
          type: string
      - name: realName
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageSysUser'
    post:
      tags:
      - 用户管理
      summary: 创建用户
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysUser'
        required: true
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultSysUser'
  /api/users/{id}:
    get:
      tags:
      - 用户管理
      summary: 获取用户详情
      operationId: getById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultSysUser'
    put:
      tags:
      - 用户管理
      summary: 更新用户
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysUser'
        required: true
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultVoid'
    delete:
      tags:
      - 用户管理
      summary: 删除用户
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultVoid'
  /api/users/{id}/reset-password:
    post:
      tags:
      - 用户管理
      summary: 重置密码
      operationId: resetPassword
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: newPassword
        in: query
        required: false
        schema:
          type: string
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultVoid'
  /api/users/{id}/status:
    put:
      tags:
      - 用户管理
      summary: 启用/禁用用户
      operationId: changeStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultVoid'
  /api/versions:
    get:
      tags:
      - 版本管理
      summary: 分页查询版本
      operationId: getList
      parameters:
      - name: pageNum
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: versionNumber
        in: query
        required: false
        schema:
          type: string
      - name: versionName
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageSysMainVersion'
    post:
      tags:
      - 版本管理
      summary: 创建版本
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysMainVersion'
        required: true
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultSysMainVersion'
...
components:
  schemas:
    ResultObject:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
    SysMainVersion:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createTime:
          type: string
          format: date-time
        createId:
          type: integer
          format: int64
        createName:
          type: string
        modifyTime:
          type: string
          format: date-time
        modifyId:
          type: integer
          format: int64
        modifyName:
          type: string
        ver:
          type: integer
          format: int32
        isDelete:
          type: integer
          format: int32
        systemName:
          type: string
          description: 主系统名称
        versionNumber:
          type: string
          description: 版本号
        fileId:
          type: integer
          description: 文件ID
          format: int64
        releaseNotes:
          type: string
          description: 发布说明
        status:
          type: integer
          description: '状态: 0-草稿, 1-已发布, 2-已下线'
          format: int32
          enum:
          - 0
          - 1
          - 2
          x-enumDescriptions:
          - 草稿
          - 已发布
          - 已下线
      description: 主系统版本
    ResultVoid:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    SysRole:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createTime:
          type: string
          format: date-time
        createId:
          type: integer
          format: int64
        createName:
          type: string
        modifyTime:
          type: string
          format: date-time
        modifyId:
          type: integer
          format: int64
        modifyName:
          type: string
        ver:
          type: integer
          format: int32
        isDelete:
          type: integer
          format: int32
        roleCode:
          type: string
          description: 角色编码
        roleName:
          type: string
          description: 角色名称
        description:
          type: string
          description: 角色描述
        status:
          type: integer
          description: '状态: 0-禁用, 1-启用'
          format: int32
          enum:
          - 0
          - 1
          x-enumDescriptions:
          - 禁用
          - 启用
      description: 系统角色
    SysUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createTime:
          type: string
          format: date-time
        createId:
          type: integer
          format: int64
        createName:
          type: string
        modifyTime:
          type: string
          format: date-time
        modifyId:
          type: integer
          format: int64
        modifyName:
          type: string
        ver:
          type: integer
          format: int32
        isDelete:
          type: integer
          format: int32
        username:
          type: string
          description: 用户名
        password:
          type: string
          description: 密码
        realName:
          type: string
          description: 真实姓名
        email:
          type: string
          description: 邮箱
        phone:
          type: string
          description: 手机号
        status:
          type: integer
          description: '状态: 0-禁用, 1-启用'
          format: int32
          enum:
          - 0
          - 1
          x-enumDescriptions:
          - 禁用
          - 启用
        lastLoginTime:
          type: string
          description: 最后登录时间
          format: date-time
        roleIds:
          type: array
          description: 角色ID列表
          items:
            type: integer
            description: 角色ID列表
            format: int64
        roles:
          type: array
          description: 角色列表
          items:
            $ref: '#/components/schemas/SysRole'
      description: 系统用户
...
    SysAccessSystem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createTime:
          type: string
          format: date-time
        createId:
          type: integer
          format: int64
        createName:
          type: string
        modifyTime:
          type: string
          format: date-time
        modifyId:
          type: integer
          format: int64
        modifyName:
          type: string
        ver:
          type: integer
          format: int32
        isDelete:
          type: integer
          format: int32
        systemName:
          type: string
          description: 系统名称
        systemUrl:
          type: string
          description: 系统URL
        appKey:
          type: string
          description: 应用Key
        appSecret:
          type: string
          description: 应用密钥
        expireTime:
          type: string
          description: 有效期截止时间
          format: date-time
        status:
          type: integer
          description: '状态: 0-禁用, 1-启用'
          format: int32
          enum:
          - 0
          - 1
          x-enumDescriptions:
          - 禁用
          - 启用
        description:
          type: string
          description: 系统描述
      description: 接入系统
    ResultSysMainVersion:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/SysMainVersion'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    ResultSysUser:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/SysUser'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
    ResultSysRole:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/SysRole'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    ResultSysAccessSystem:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/SysAccessSystem'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    OrderItem:
      type: object
      properties:
        column:
          type: string
        asc:
          type: boolean
    PageSysMainVersion:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/SysMainVersion'
        total:
          type: integer
          format: int64
        size:
          type: integer
          format: int64
        current:
          type: integer
          format: int64
        orders:
          type: array
          writeOnly: true
          items:
            $ref: '#/components/schemas/OrderItem'
        optimizeCountSql:
          $ref: '#/components/schemas/PageSysMainVersion'
        searchCount:
          $ref: '#/components/schemas/PageSysMainVersion'
        optimizeJoinOfCountSql:
          type: boolean
          writeOnly: true
        maxLimit:
          type: integer
          format: int64
          writeOnly: true
        countId:
          type: string
          writeOnly: true
        pages:
          type: integer
          format: int64
    ResultPageSysMainVersion:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/PageSysMainVersion'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    PageSysUser:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/SysUser'
        total:
          type: integer
          format: int64
        size:
          type: integer
          format: int64
        current:
          type: integer
          format: int64
        orders:
          type: array
          writeOnly: true
          items:
            $ref: '#/components/schemas/OrderItem'
        optimizeCountSql:
          $ref: '#/components/schemas/PageSysUser'
        searchCount:
          $ref: '#/components/schemas/PageSysUser'
        optimizeJoinOfCountSql:
          type: boolean
          writeOnly: true
        maxLimit:
          type: integer
          format: int64
          writeOnly: true
        countId:
          type: string
          writeOnly: true
        pages:
          type: integer
          format: int64
    ResultPageSysUser:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/PageSysUser'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
    PageSysRole:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/SysRole'
        total:
          type: integer
          format: int64
        size:
          type: integer
          format: int64
        current:
          type: integer
          format: int64
        orders:
          type: array
          writeOnly: true
          items:
            $ref: '#/components/schemas/OrderItem'
        optimizeCountSql:
          $ref: '#/components/schemas/PageSysRole'
        searchCount:
          $ref: '#/components/schemas/PageSysRole'
        optimizeJoinOfCountSql:
          type: boolean
          writeOnly: true
        maxLimit:
          type: integer
          format: int64
          writeOnly: true
        countId:
          type: string
          writeOnly: true
        pages:
          type: integer
          format: int64
    ResultPageSysRole:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/PageSysRole'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    PageSysAccessSystem:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/SysAccessSystem'
        total:
          type: integer
          format: int64
        size:
          type: integer
          format: int64
        current:
          type: integer
          format: int64
        orders:
          type: array
          writeOnly: true
          items:
            $ref: '#/components/schemas/OrderItem'
        optimizeCountSql:
          $ref: '#/components/schemas/PageSysAccessSystem'
        searchCount:
          $ref: '#/components/schemas/PageSysAccessSystem'
        optimizeJoinOfCountSql:
          type: boolean
          writeOnly: true
        maxLimit:
          type: integer
          format: int64
          writeOnly: true
        countId:
          type: string
          writeOnly: true
        pages:
          type: integer
          format: int64
    ResultPageSysAccessSystem:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/PageSysAccessSystem'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    PageSysAccessLog:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/SysAccessLog'
        total:
          type: integer
          format: int64
        size:
          type: integer
          format: int64
        current:
          type: integer
          format: int64
        orders:
          type: array
          writeOnly: true
          items:
            $ref: '#/components/schemas/OrderItem'
        optimizeCountSql:
          $ref: '#/components/schemas/PageSysAccessLog'
        searchCount:
          $ref: '#/components/schemas/PageSysAccessLog'
        optimizeJoinOfCountSql:
          type: boolean
          writeOnly: true
        maxLimit:
          type: integer
          format: int64
          writeOnly: true
        countId:
          type: string
          writeOnly: true
        pages:
          type: integer
          format: int64
    ResultPageSysAccessLog:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/PageSysAccessLog'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
    SysAccessLog:
      type: object
      properties:
        id:
          type: integer
          description: 主键ID
          format: int64
        traceId:
          type: string
          description: 链路追踪ID
        serviceName:
          type: string
          description: 服务名称
        logType:
          type: string
          description: 日志类型
        httpMethod:
          type: string
          description: HTTP方法
        requestUri:
          type: string
          description: 请求URI
        clientIp:
          type: string
          description: 客户端IP
        statusCode:
          type: integer
          description: 响应状态码
          format: int32
        successFlag:
          type: integer
          description: 是否成功
          format: int32
          enum:
          - 0
          - 1
          x-enumDescriptions:
          - 失败
          - 成功
        elapsedMs:
          type: integer
          description: 耗时毫秒
          format: int64
        tokenType:
          type: string
          description: Token类型；令牌类型
          enum:
          - INTERNAL
          - EXTERNAL
          x-enumDescriptions:
          - 内部用户令牌
          - 外部系统令牌
        appKey:
          type: string
          description: 应用Key
        userId:
          type: string
          description: 用户ID
        userName:
          type: string
          description: 用户名称
        enterpriseId:
          type: string
          description: 企业ID
        enterpriseName:
          type: string
          description: 企业名称
        bizType:
          type: string
          description: 业务类型
        bizId:
          type: string
          description: 业务ID
        projectId:
          type: string
          description: 项目ID
        tenderId:
          type: string
          description: 标段ID
        fileId:
          type: string
          description: 文件ID
        fileName:
          type: string
          description: 文件名称
        requestHeaders:
          type: string
          description: 请求头摘要
        requestBody:
          type: string
          description: 请求体摘要
        responseBody:
          type: string
          description: 响应体摘要
        errorType:
          type: string
          description: 异常类型
        errorMessage:
          type: string
          description: 异常信息
        createTime:
          type: string
          description: 记录时间
          format: date-time
      description: 系统访问日志